requirements.txt
numpy
regex
chromadb
pydantic
onnxruntime
safetensors
huggingface-hub
bm25s
fasttext-wheel
